add source maps, modules, and comma formatter
