+++ parser/formatter.js
@@ -1,0 +1,9 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+function formatComma(input) {
+    let output = "";
+    console.log(input);
+    return output;
+}
+exports.formatComma = formatComma;
+//# sourceMappingURL=formatter.js.map
