+++ parser/parser.ts
@@ -17,14 +17,9 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
+import { formatComma } from "./formatter";
 
-function formatComma(input: string): string {
-    let output: string = ""
-    console.log(input)
-    return output
-}
-
-function parseJsonss(styles: object, pretty: boolean, debug: boolean): string {
+function parseJsonss(styles: Object, pretty: boolean, debug: boolean): string {
     let output: string = ""
     for (const [key, value] of Object.entries(styles)) {
         if (debug) {console.log("🔎 parsing", key, "=", value)}
@@ -34,7 +29,7 @@ function parseJsonss(styles: object, pretty: boolean, debug: boolean): string {
         if (pretty) {
             if (debug) {console.log("…formatting", value, "pretty print =", true)}
 
-            val = formatComma(
+            val = formatComma (
                 JSON.stringify(value)
                     .replace(/_/g, "-")
                     .replace(/"/g, "")
@@ -49,7 +44,7 @@ function parseJsonss(styles: object, pretty: boolean, debug: boolean): string {
         } else {
             if (debug) {console.log("…formatting", value, "pretty print = ", false)}
 
-            val = formatComma(
+            val = formatComma (
                 JSON.stringify(value)
                     .replace(/_/g, "-")
                     .replace(/"/g, "")
